backend/requirements.txt
# Django Core
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.8.0

